Cleaned up commits and code + fixed some smaller stuff
Added caching mechanism, like this the plugin can be run in a
production environment (if desired) without slowing down the page.

The changes speed the plugin up significantly (~60 files takes about
1.8 seconds on every page rendering, with new code it's down to 0.04s
and if something is changed around 1s)

Check for usergroups before running the plugin. Prevents it from
executing when visitors watch the page and when manager users with no
rights to change element files reload the pages

Added debugging possibilites including speed measurement.

Cleaned up the commit mess and mad one out of 6 =)…sorry, that was just
pasted in directly on github when adding the features^^